@@ -1,6 +1,6 @@
 {
   "name": "conseiljs",
-  "version": "5.0.6-1",
+  "version": "5.0.6-2",
   "description": "Client-side library for Tezos dApp development.",
   "browser": "dist/index-web.js",
   "main": "dist/index.js",
@@ -70,14 +70,14 @@
     "big-integer": "1.6.48",
     "blakejs": "1.1.0",
     "bs58check": "2.1.2",
-    "jsonpath-plus": "4.0.0",
+    "jsonpath-plus": "5.0.2",
     "moo": "0.5.0",
     "nearley": "2.19.1"
   },
   "devDependencies": {
-    "@types/chai": "4.2.12",
+    "@types/chai": "4.2.14",
     "@types/chai-as-promised": "7.1.3",
-    "@types/mocha": "8.0.3",
+    "@types/mocha": "8.2.0",
     "@types/nock": "11.1.0",
     "@types/node": "14.0.13",
     "@types/node-fetch": "2.5.7",
@@ -88,18 +88,18 @@
     "conseiljs-softsigner": "5.0.3",
     "copyfiles": "2.3.0",
     "coveralls": "3.1.0",
-    "eslint": "7.9.0",
+    "eslint": "7.18.0",
     "glob": "7.1.6",
-    "loglevel": "1.7.0",
-    "mocha": "8.1.3",
-    "nock": "13.0.4",
+    "loglevel": "1.7.1",
+    "mocha": "8.2.1",
+    "nock": "13.0.6",
     "node-fetch": "2.6.1",
     "nyc": "15.1.0",
     "terser-webpack-plugin": "3.0.8",
     "ts-node": "8.10.2",
     "tsconfig-paths-webpack-plugin": "3.3.0",
-    "typedoc": "0.19.2",
-    "typedoc-plugin-markdown": "3.0.3",
+    "typedoc": "0.18.0",
+    "typedoc-plugin-markdown": "2.4.2",
     "typescript": "3.8.3",
     "webpack": "4.44.2",
     "webpack-cli": "3.3.12"
